+++ tests/build_dir.rs
@@ -6,16 +6,16 @@ mod util;
 use util::{copy_of_testdata, run};
 
 #[test]
-fn gitignore_respected_in_copy_by_default() {
+fn gitignore_respected_when_enabled() {
     // Make a tree with a (dumb) gitignore that excludes the source file; when you copy it
-    // to a build directory, the source file should not be there and so the check will fail.
+    // to a build directory with gitignore enabled, the source file should not be there and so the check will fail.
     let tmp = copy_of_testdata("factorial");
     // There must be something that looks like a `.git` dir, otherwise we don't read
     // `.gitignore` files.
     create_dir(tmp.path().join(".git")).unwrap();
     write(tmp.path().join(".gitignore"), b"src\n").unwrap();
     run()
-        .args(["mutants", "--check", "-d"])
+        .args(["mutants", "--check", "--gitignore=true", "-d"])
         .arg(tmp.path())
         .assert()
         .stdout(predicates::str::contains("can't find `factorial` bin"))
@@ -30,6 +30,22 @@ fn gitignore_can_be_turned_off() {
     write(tmp.path().join(".gitignore"), b"src\n").unwrap();
     run()
         .args(["mutants", "--check", "--gitignore=false", "-d"])
+        .arg(tmp.path())
+        .assert()
+        .success();
+}
+
+#[test]
+fn gitignore_not_respected_by_default() {
+    // Make a tree with a (dumb) gitignore that excludes the source file; when you copy it
+    // to a build directory by default (gitignore=false), it should succeed because gitignore is ignored.
+    let tmp = copy_of_testdata("factorial");
+    // There must be something that looks like a `.git` dir, otherwise we don't read
+    // `.gitignore` files anyway.
+    create_dir(tmp.path().join(".git")).unwrap();
+    write(tmp.path().join(".gitignore"), b"src\n").unwrap();
+    run()
+        .args(["mutants", "--check", "-d"])
         .arg(tmp.path())
         .assert()
         .success();